requirements.test2.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.1
bidict==0.23.1
billiard==4.2.1
blinker==1.8.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.1
django-allauth==64.2.1
django-channels==0.7.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
exceptiongroup==1.2.2
gevent==24.2.1
gevent-socketio==0.2.1
gevent-websocket==0.10.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.0.3
groq==0.11.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.4.2
MarkupSafe==2.1.5
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.2
pandas==2.2.3
pillow==10.4.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
py-dotenv==0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.1.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sphinx-me==0.3
sqlparse==0.5.1
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.2
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.3
